base user api route controller
