+++ flask_server/src/__init__.py
@@ -1,2 +1,1 @@
-from .server import Server as Server
-from src.routes import Router as Router
+
