+++ flask_server/src/controllers/__init__.py
@@ -1,0 +1,1 @@
+from .api_route_controller import api_controller as api_route_controller
