+++ flask_server/src/routes/__init__.py
@@ -1,1 +1,1 @@
-from .router import Router as Router
+
